feat(opilab-demo): support "exec" to opicluster-master-{1,2,3} node
It is useful to be able to access the master nodes conveniently. Support
that via

  ./demo.sh exec opicluster-master-1 hostname

Also, some related cleanups to make this easier.

Also, extend `./demo.sh ssh_copy_id ... ` to support the master nodes.
But actually, we usually don't need to install the SSH key on all nodes,
because the exec command anyway proxies through tgen1 host.

Signed-off-by: Thomas Haller <thaller@redhat.com>