feat(opilab-demo): add environment variables for IP addresses
The demo currently only works with the IPU inside dh4 host. It will be
extended to also work with the Marvell DPU in dh3.

Also, it will be useful to use the script in other clusters. For that,
add environment variables. For one, this gives (variable) names instead
of repeating the same IP address throughout the code. But also, this
makes the script configurable to the caller, who can set the environment
variables.

Signed-off-by: Thomas Haller <thaller@redhat.com>