feat(opilab-demo): extend "demo.sh" script for Marvell DPU and external cluster
We want that the "demo.sh" script supports multiple use cases:

- the two cluster design (where microshift runs on the DPU) with one
  OCP cluster and several DPUs (IPU and Marvell DPU) on separate nodes.
  While current DPU operator may have issues with multiple DPUs,
  we should be able to still `./demo.sh redploy $KIND` (which would
  deconfigure the operator from one node and configure it on the other).

- in the future, we want instead to use a one cluster design where the
  DPU is a node in the OCP cluster. This would then also support
  Phantomlake.

- we also want to be able to use the "demo.sh" in other clusters, outside
  the OPI lab. The benefit is that the similar demo setup can easily be
  created on other systems.

This makes the script more versatile, but also more complicated. The
latter is a problem, because the script should still be reviewable and
understandable. On the other hand, to demo multiple DPUs and migrating
between them is an important part. It is also important to use the
script in other clusters, without having to fork the script. So, the
script can do more now, which makes it more complicated (and is
undesirable). But overall, it is more important that we can do those
additional things.

Signed-off-by: Thomas Haller <thaller@redhat.com>